fix: Use contact email as primary email address
The change at https://github.com/uktrade/staff-sso/pull/332 removed the
contact email address from the emails array, so we're showing the
incorrect email for some people.

Note: the emails array can be empty in some cases, and so this could
raise an exception. However, think this should be fixed on SSO's side,
at least for now.