feat: bump SSRI
This is due to a vulnrabiliy. Slightly unfortunately this is a
transitive dependency, and the intermeiate dependencies seem to depend
on earlier thatn the fixed version. Have only done a brief bit of
testing, but looking in the changlog of SSRI, while there _have_ been
breaking changes, SSRI's API hasn't changed completely, so I think a
reasonable chance the breaking bits weren't being used.