refactor: create permissions in post_migrate hook
This is so in Django 3.2 the permissions are added to the group after
they are created. In Django 3.2, doing this in the migration after the
signal, didn't find the permission instances needed added to the group.

Suspect that the AppConfig instance in datasets wasn't actually used
before: the default_app_config is so that Django (and certainly
pytest-django) finds it and loads it as required.