feat: avoid warning in RStudio about timedatectl and systemd
When connecting to the database, suspect in R was calling
Sys.timezone(), which in turn was calling, timedatectl which output the
warning:

System has not been booted with systemd as init system (PID 1). Can't operate.
Failed to create bus connection: Host is down
Warning message:
In system("timedatectl", intern = TRUE) :
  running command 'timedatectl' had status 1

Manually setting the timezone seems to avoid this.