feat: bump Quicksight SDK to try to avoid PDF filtering bug
Embedded dashboards have a download as PDF feature, which _should_
preserve the state of any filtering chosen by the user: they do when the
dashboard is accessed directly, but not when embedded.

This is a very speculative change: nothing in the SDK changelog suggests
this bug is fixed (although I don't really understand everything in the
changelog)

Keeping the old SDK file so requests during deployment still work.