feat: tie sessions to group of IP addresses
This is to mitigate the risk of session hijacking. Ideally, we would tie
sessions to a single IP, but we can't prevent users coming from multiple IP
addresses in some cases. Specifically for subdomains, so for tools and
visualisations, users can consistently come from a different IP to the root
domain. This causes problems for the SSO flow on a subdomain, since it involves
a redirection to the root domain which needs the session to be consistent
between it and the subdomain to complete the flow securely.

This deliberately supports both APPLICATION_IP_WHITELIST and
IP_ALLOWLIST_GROUPS for now for a rollout without breaking anything, although
APPLICATION_IP_WHITELIST will be removed later.