refactor: remove conditions that are unlikely to happen
In each of these cases, we know we have at least one dataset of this type since
we found it earlier in the request. Since this is not in a REPEATABLE READ
transction, we could technically hit errors if we hard delete a dataset
mid-request, but this is so rare - we would have to manually do it in the
database. I think the combination of this being rare, the fact that it's not a
biggie even if it happens, means it's better to have simpler code which we can
optimise more easily