docs: remove todos
We may or may not refactor the search - we might not have a need. Better to
remove in case we don't.

Also I think the `_get_tags_as_dict` seems very reasonable already - gets all
the tags in one query which avoid the N+1 problem for tags. Potentially we get
some tags we don't need, but I don't think we expect to have many.