fix: attempt to do gevent monkey patching in Celery worker that doesn't run start.py
The change to the download view may just be due to the test environment - it
was causing errors in tests:

> execute cannot be used while an asynchronous query is underway

At most as I can tell, it was something to do with the fact that there was a
connection to the admin db from inside the manually spawned greenlet that
queries the datsets db. This feels slightly messy anyway, so I think it's a
step forward to only query the datasets db from the greenlet.