feat: search by popularity as flag rather than switch
Converting the search by popularity to a flag to allow specific users (such as
member of the Data Infrastructure team) to test it in production before we roll
it out to all users.

Also, putting it at the bottom of the list, so default behaviour is as it is
for all users.