feat:add new search_vectors for name,short_desc,tags,desc & use in sort
This adds an individual search vector for each of name, short_description, tags and description for use in search ranking when sorting by relevance. Whether a dataset is bookmarked or not and if the source table name is an exact match is prioritised over the search vector fields when sorting by relevance.