feat: stronger normalisation and no cover density in dataset title search rank
When searching for "companies in the uk", the dataset "Companies in the UK" was
far below other result. Specifically, below those of the form "UK Companies xxx
xxxx xxxx". Suspect this is due to

- The fact that cover density strongly prioritises results with the words "UK"
  and "Companies" close together.

- The fact that normalization=1, dividing the search rank by log of the
  document length, doesn't penalise longer titles enough.

So trying out turning off cover density matching for the name, and setting
normalization=2, dividing the search rank by the document length.

Can't be completely sure how much this helps before it hits production data,
but it's behind a feature flag.