feature: allow user to cancel query in data explorer
This gets the query pid when a query is executed and saves it against the QueryLog model. The user is then presented with a cancel button which makes an ajax request to cancel the query using pg_cancel_backend. The green spinner then handles the exception thrown in the async task and prints it out verbatim as it's a decent error message: "canceling statement due to user request"