feature: allow user to toggle bookmarks from the search results page
This adds a listener to the bookmark icons and make an ajax requests to new
endpoints. The icon is toggled in javascript before the ajax call is made and
some refactoring was done to the template to make this easier.

In terms of UI/robustness, kept it simple in terms of not worrying about
retrying in the case of failures, or showing the user the result of any such
failures, and not giving any indication that there is a request in progress
that takes time or can fail. Justified since it's not critical that the server
is absolutely in sync with the browser for bookmarks.

That being said, some small choices were made to increase robustness. Rather
than using the existing toggle endpoint that flips state no matter what it is,
using new endpoints that explicitly set a bookmarked/not bookmarked state -
this reduces the impact of a single failure in a string of requests. And a
simple promise-based queue means requests can't overtake each other, which
means there is a smaller chance that the server will become out of sync with
the browser.

The removal of listening to the "change" event on the search box - this was to
prevent searches being done after:

- the user clicking into the search box (and maybe performing a search)
- then the user clicking on the bookmark button on a search result

which was a bit strange. This also means that searches are no longer done when
tabbing out of the search box, but this actually feels like an improvement - it
was a bit strange before, and inconsistent with both GOV.UK and Google.