feat: use `values` to better avoid N+1 issues when querying ReferenceDataset in bulk
Looking at APM, we still have N+1 queries in Data Explorer in some cases for
ReferenceDataset. Suspect this is due to code it has in its __init__ method.

We could do more "selected_related" and/or "prefetch_related", but it's much
easier, and robust, and likely quicker, to avoid these and call "values" that
doesn't run __init__ at all because model instances are not constructed.

Note that `Credentials-Db-Id` isn't used by Superset, and so can be removed.