feat: fix multipart upload in local dev environment
The etag header is needed for multipart uploads, but it wasn't configured to be
accessible from Javascript, CORS-wise.

There is a localstack environment variable, EXTRA_CORS_EXPOSE_HEADERS, but that
didn't seem to work - it only returned the header names in
Access-Control-Expose-Headers in response to the preflight request, not the
request proper where it needs to to be.

Suspect that there is some conflict between EXTRA_CORS_EXPOSE_HEADERS and the
"proper" CORS config for S3, where the S3 config takes priority for certain
requests.