fix: bugs when deleting multiple files
Deletes were happening in a strange order or not at all - so while I don't
really know what the problem was, from testing locally this version is less
buggy, and the code is simpler since it doesn't attempt to do them in parallel.

The deletes are still batched in batches of 1000, so it shouldn't make it too
much slower. And the previous delete code in the Angular JS Your files didn't
do parallel deletes, so it shouldn't be a regression in terms of performance.

DT-771