feat add line after every collections item
It's potentially slightly odd to be re-using the "search-result" class, but I
think it's reasoanble to use the same class as the front page to get as much
consistency between those results and these items. The worst thing may be the
name "search-result" since it's now also used not as a search result, but
leaving as is. This commit is still a step forward in terms of UX consistency
and code consistency between different parts of the code.

DT-807