fix: force s3ssync to either not make AAAA requests, or ignore failures
It looks like s3sync wasn't responding well to AAAA request failures from our
DNS Rewrite proxy. Rather than supporting AAAA, trying best to not use ipv6 to
KISS (since these is all internal).

The main thing that I think makes this more robust is the move to Debian and
trying to get it to prefer A records.  Although I still see requests in the DNS
rewrite proxy for AAAA records, s3sync doesn't seem to be going wrong in
response to them. While it's not great to not completely understand everything
that's going on, might have to live with the fact that this seems to be working
now.

Requires merge and release of https://github.com/uktrade/mobius3/pull/50