tests: make tests less flakey by moving them out of UK summer time
The test_create_chart function for some reason can cause some tests in
test_admin.py to fail, since if test_create_chart runs, then the times on these
assertions end up an our off. Moving these times out of UK summer time is a bit
of a guess, but it seems to fix the issues. Potentially the test_create_chart
function causes some setting in the application or the database that persists
to the next test? Not sure.

This wasn't found until we started investigating parallelising tests, where
different sets of tests would run on different instances in different orders.

Potentially this commit is hiding a production issue to do with daylight
savings/time zones, but I think we're just going to have to live with it - some
things being an hour off, for our use cases, we can live with

DT-270