feat: login to environment-specific namespace on Quicksight
There is still more to be done to fully separate environments, this is just the
first stage. Some environment variables were manually specified in
per-environment JSON - this change is tidying them up and putting them in the
common JSON so there is a better chance that each environment is setup
correctly.

DT-956