feat: replace user email with ssoid to hide reader accounts when sharing dashboards
There is often confusion when sharing quicksight dashboards due to the seperate READER and AUTHOR users. While we would like to have a single user for these, in the meantime this PR replaces the {user.email} in the name of READER users with their {sso_id}. This should hopefully hide these fromthe options when sharing a dashboard.

This includes a fix to correct the embed role for non prod environments,
an oversight from previous changes to quicksight namespaces