Increased the max number of db connections available for quicksight celery tasks (#2430)
After some investigation around permission syncs and other tasks that hinted
that tasks weren't running due to contention for a limited resource. This is a
speculative change that increases the limit for db connections for those tasks.
We are deliberately not increasing limit everywhere to avoid hitting the max
number of connections on the db itself.