feat: avoid running more than one send_notification_emails task at a time
The job is set to run every 5 minutes, but looks like the job can in some cases
take more than 5 minutes. Not really sure how long the "timeout" value should
be - this is the maximum time the lock can be held. Maybe this is too short.
However, it's still longer than now which is essentially "0" beceause there is
no lock

Suspect this could be a contributing factor to various background jobs not
running - if many send_notification_emails jobs were running, for various
reasons they could have blocked otherb.

DT-1162