build(deps): use SQLAlchemy's database connection pool
Having a look at what was being used before,
https://github.com/jneight/django-db-geventpool, suspect there are a few race
conditions in the database connection pool code. We could fork (and we did try
that), but

- It felt a bit like reinventing the wheel
- It used very little gevent-specific code

So, opting to use https://github.com/altairbow/django-db-connection-pool, a
thin layer over SQLAlchemy's database connection pooling code, with the hope
that with SQLAlchemy's popularity, it would address these race conditions, and
be generally more robust. Note that because of gevent's monkey patching, using
gevent with the pool _should_ work fine.

This change includes moving from psycopg2-binary to psycopg2 since
https://github.com/altairbow/django-db-connection-pool requires that. It
appears that for production use, psycopg2 is recommended. The downside of that
is that it needs more dependencies to compile.

Another downside is that this no longer has a logger.error when all connections
has been allocated. However, I wonder if this is a better longer term bet -
more battle tested pooling code without our own fork of pooling code which is
what the logger.error came from.

However, the default timeout that a request will wait before getting a
connection is 30 seconds - compared to what I suspect was no timeout before,
other than the 60 second timeout before a user gets a 50x from either nginx or
the load balancer.  This admittedly may introduce some user-facing pain
de-facto 60 second timeout to 30 seconds. But - it will force the situation to
be addressed. Users should not be waiting 30 seconds for a database connection.