feat: if no match for an email exists widen the search
This is to work around the fact that users can have either @trade or @businessandtrade emails currently. We are at whim of the SSO service as to which one a user has. This change allows us to search for a @trade user but if no user is found, return an @businessandtrade user