feat: update datasets database terraform with what's in prod
This updates the terraform for the datasets database with the config that's in
production. Suspect there were changes made in the console (from PostgreSQL
version upgrades and/or responses to pen test)

So this is maybe more of a "refactor" in terms of production - shouldn't
actually result in any changes.