feat: update datasets database terraform with what's in prod part 2
This updates the terraform for the datasets database with what _should_ be in
prod.

The change in 4bd9affa354e21c62e4ba8a9be805c11cf22e1d8 attempted to do this,
but for some reason wasn't quite right and changes still tried to be made. This
is attempt 2