feat: allow users to list the mirror bucket
This gives users in Data Workspace the ability to list the contents of the
mirror bucket so they can see what's been mirrored, .e.g. from Python scripts
inside tools.

At the moment however, I suspect for a user to be able to use this, they would
have to have the Tools access IAM role arn manually cleared in their user
record, which would force their IAM role to be reconstructed with the new
permission. I'm not sure of a way to do this in bulk right now.