build: add latest QuickSight embedding SDK (but don't use it)
This adds the latest QuickSight embedding SDK file. However, it doesn't yet use
it. This is to make sure it's fully deployed so in the deployment when we _do_
use it, there shouldn't be broken requests by the browser expecting a file, but
the server it hits doesn't have it.