feat: use latest Quicksight embedding SDK
This updates the version of the Quicksight embedding SDK to the latest, which
includes some breaking changes and so we also change how we use it. But
hopefully resulting in the same/very similar behaviour.

This is done to hopefully fix issues with external links in QuickSight
dashboards. They used to to work, but now they seem to open blank tabs.
Invgestigating the dev console showed there is some issue to do with a null
"opener", which suggests potentially the old SDK is hitting on a security
barrier that is now present for some reason.