feat: set COOP to same-origin-allow-popups for Quicksight
::Description of change

This overrides the Cross-Origin-Opener-Policy HTTP header from its Django
default of 'same-origin' to 'same-origin-allow-popups' for the pages where
Quicksight dashboards are embedded.

This is done to fix the issue that QuickSight embedded dashboards cannot open
new tabs. Or more specifically, they can open new tabs, but they remain blank.
This fix is done at the suggestion of Quicksight via a support ticket.

From testing this on staging, it does fix the issue. However, before getting it
to production we should be sure why it not working, why it broke now, why this
fixes it, and be sure that the fix is safe - as in, we should know it doesn't
meaningfully introduce more risk.

::Why it's not working

From looking around its Javascript source, inside its embedded iframe,
Quicksight appears to be using the technique from
https://stackoverflow.com/a/40593743/1319998 to open popups:

```javascript
var yourWindow = window.open();
yourWindow.opener = null;
yourWindow.location = "http://someurl.here";
```

This is actually The Right Thing to prevent the page in the popup/tab from
reaching _back_ to the Quicksight page via its own window.opener property (at
least, The Right Thing that supports older browsers, because I think the
'noopener' parameter does the same thing but works only on newer browsers). But
it unfortunately can't be done from

- a cross domain iframe - like an embedded Quicksight dashboard
- where the parent page has  "Cross-Origin-Opener-Policy: same-origin" - like
  Data Workspace

because in this case `window.open` returns `null`, and so location (and opener)
can't be set on it. Why it returns `null` in this case - it's in the spec. In
this exact situation it makes `window.open` behave as though you passed
`noopener` to it, which in turn makes window.open return null.

In more detail, from [The rules for choosing a
navigable](https://html.spec.whatwg.org/multipage/document-sequences.html#the-rules-for-choosing-a-navigable)

> 3. If currentDocument's cross-origin opener policy's value is "same-origin"
> or "same-origin-plus-COEP", and currentDocument's origin is not same origin
> with currentDocument's relevant settings object's top-level origin, then: 1.
> Set noopener to true.  2. Set name to "_blank".  3. Set windowType to "new
> with no opener".

and then from [7.2.2.1 Opening and closing
windows](https://html.spec.whatwg.org/multipage/nav-history-apis.html#apis-for-creating-and-navigating-browsing-contexts-by-name),
when it's describing what window.open does

> 14. If noopener is true or windowType is "new with no opener", then return
> null.

::Why it broke now

The HTTP header "Cross-Origin-Opener-Policy: same-origin" was added to Django 4
and not present earlier
https://github.com/django/django/commit/db5b75f10fe211af9fab9094f937436760db8488,
and we only recently moved to moved to Django 4 in
https://github.com/uktrade/data-workspace/commit/553f016307c764ae9fa26a26c9ffbdd2dee48387.

::Why this fixes it

"Cross-Origin-Opener-Policy: same-origin-allow-popups" makes all popups opened
by window.open return a non-null object to Quicksight's code, and allows it to
continue to set its location (and null opener).

"Cross-Origin-Opener-Policy: same-origin" affects cross origin access to Data
Workspace in 2 ways - if Data Workspace is opened by a cross origin site, or if
Data Workspace itself opens up a cross origin site.

1. Data Workspace is opened by a cross origin site

   The behaviour of "Cross-Origin-Opener-Policy: same-origin" should be exactly
   the same as "Cross-Origin-Opener-Policy: same-origin-allow-popups" for the case
   of Data Workspace being opened by another site.

   From https://developer.chrome.com/blog/coop-restrict-properties
   > Set **same-origin-allow-popups**, which blocks all cross-origin
   > interactions that open your site in a popup

   and from

   https://html.spec.whatwg.org/dev/browsers.html
   > **"same-origin-allow-popups"**
   >
   > This forces the creation of a new top-level browsing context for the
   > document, unless its predecessor specified the same cross-origin opener
   > policy and they are same origin.

2. Data Workspace opening up other sites

   Because of the weaker separation between domains there is an increased risk
   of Spectre-style attacks. However, browsers respond to
   "Cross-Origin-Opener-Policy: same-origin-allow-popups" by disabling certain
   Javascript APIs that are used in these attacks, e.g. high precision timers. So
   although the separation is less, there is an automatic mitigation. For
   completeness I suspect we already didn't have the high precision timers and
   other Spectre-related APIs enabled because we didn't, and still don't, use the
   Cross-Origin-Embedder-Policy that is apparently required to enable what I think
   it termed "cross-origin isolation" that allows these APIs.

   In terms of more direct Javascript-based access, the site in the popup can
   also not reach back to Data Workspace because Quicksight sets `window.opener =
   null`. The change also should only affect the pages that show Quicksight
   dashboards, so even if somewhere else in Data Workspace doesn't open popups in
   this way, this change shouldn't decrease protections there.

   This change shouldn't affect the separation between Data Workspace and
   Quicksight itself - that is handled with Cross-Origin-Embedder-Policy, our CSP,
   and Quicksight's CSP: all of which are unchanged. Also from
   https://docs.google.com/document/d/1zDlfvfTJ_9e8Jdc8ehuV4zMEu9ySMCiTGMS9y0GU92k/edit?pli=1

   > Because COOP is defined in terms of browsing context groups, it doesn't
   > apply to iframes; the browser ignores COOP on documents which aren't
   > top-level, and allows iframes to access their cross-origin ancestors even
   > if these ancestors set COOP.

   So however secure/insecure the separation was before, it's just as it was