fix: Catch `None` hostname when calculating metrics for app instance
If a tool has failed to start, there is a chance the hostname
was not set. If you then try to view the details of that tool
in the admin you get a 500 error as it expects the hostname
to not be None.

Workaround this by raising an "expected" exception