refactor: move to Terraform AWS Provider 5.35.0
This includes the changes required to move the Terraform AWS provider to
5.35.0. This is done because we're trying to make it easier to apply, and don't
want to waste time solving rough edges that will have to be changed anyway when
we eventually upgrade.

Note that when applying the terraform you will probably have to run `terraform
init -upgrade` to allow selection of the new version.