refactor: move from template_file to templatefile
This changes one use of the deprecated template_file provider to the now built
in templatefile function in terraform.

This is done because the template_file provider as no current version for the
arm processor (e.g. modern macs), and so it should make it easier to apply the
terraform.