feat: only apply cloudwatch subscription filter if configured to do so
This only creates a aws_cloudwatch_log_subscription_filter on the mirrors sync
job if configured to do so. This is consistent with the other cases of
aws_cloudwatch_log_subscription_filter

(Note we might not even be using the mirror sync job any more? But leaving
investigating that to another time)