fix: bootstrappable image tags for task definitions
Starting up a new Data Workspace environment results in task definitions
refering to Docker containers with a tag, so ending in just ":", which is not
valid.

Not sure on the best default, but "latest" is probably reasonable, because that
usually (always?) exists once there is an image in the ECR repository.