Increasing the number of admin instances from 2 to 3, for prod.
This is because there was high cpu usages for users, therefore we added more
instances to mitigate this.

It should be safe as each instance will use no more than 37 data base
connections and even during a deployment, we will never have more than 10
instances including celery so we will never use more that 370 data base
connections, which is less than the 403 which the data base supports.