refactor: move out tools-supporting Dockerfiles for metrics and s3sync
This is more of a cross-repo refactor: moving things in from
https://github.com/uktrade/data-workspace-frontend to a more appropriate place.

Note that the metrics and s3sync containers are manually built and released:
there is no Jenkins build or anything like that. While that's not great and we
should sort that, at least that means this doesn't break anything.

The other side of this is at
https://github.com/uktrade/data-workspace-tools/pull/96