feat: remove some of the locking when granting to a role
This removes what is thought to be unnecessary locking - although we do need
locking when GRANTing SELECT to a table, because the permissions are all stored
in one field in pg_class, GRANTing a role to another role is done by adding a
row to pg_auth_members, which shouldn't need locking.

This is done as a small step towards hopefully more robust tools launches -
while this change might look like it decreases robustness, locking adds to
slowness, and if it turns out we _do_ need this locking, we need to known now,
as we're about to build on the pattern that assumes that changes to role
membership doesn't need locking.

There are existing integration tests that cover at least the main parts of
tools being granted enough permissions to access datasets. While not impossible
to make a test that asserts that the locking removed here is indeed
unnecessary, think it will take quite a long time and become not worth it.