feat: remove Redis-based locking when making database permissions
This removes the Redis-based locking when creating database users (and then
when cleaning them up).

This is to remove the case where Redis locks are lost due to their limited
lifetimes (which are in place so if the server goes down 1/2 way through, they
are not kep indefinitely), which can cause "tuple concurrently updated" errors.
It's also to simplify the setup - only PostgreSQL is involved and has to be
reasoned about.

This is a stepping stone towards a setup where each table has a permanent role
where it's the only role that's granted permissions to the table to work around
limits on how many users can be granted select to a table.

We could potentially have more granular locking by using different lock IDs.
However, we already weren't using granular locking, so keeping the change as
small as possible.