feat: wrap setting settings on database users/roles in a lock
It looks like in addition to PostgreSQL not handling concurrent changes to
permissions on database objects, it also doesn't handle ALTER USER ... SET ...
concurrently well. This now wraps that code in a lock.

This shouldn't meaningfully increase non-parallelism _too_ much, because there
the same lock needs to be grabbed in all cases anyway.