feat: grant database connect via role to avoid hitting limits
This is a small step of the plan where temporary database users / user
permanent roles are never granted privileges directly on various objects,
because they have small limits (their privileges have to squeeze into 1
database page, ~8kb).

Instead, they will all be granted membership of intermediate roles, and only
those roles will be granted permissions to the various objects.

The objects in question are databases (where CONNECT is needed by the user's
temporary user), schemas (where USAGE is needed by permanent roles) and tables
(where SELECT is needed).

This change handles just the database case.

Note: the name of the role is not arbitrary - it has the OID of the object
being granted privileges on, in this case the database. The alternative is the
human readable name. While that might seem better for some debugging, it runs
the risk of ending up with a mismatch if things are renamed - permissions
follow objects if they're renamed. While the database is not likely to be
renamed, tables might be renamed, including just manually in the database.
Suspect it'll be more confusing if we end up with some roles given permissions
to tables that don't match their name. Confusing, and therefor a security risk.

Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Nick Ross <nick.ir.ross@gmail.com>