feat: grant privileges on schemas via roles to avoid limits when starting tools
This avoids granting USAGE on schemas directly to users. Instead, it grants via
a per-schema role. This is to work around the de-facto limits that various
database objects have on how many grants they can have.

This avoids granting USAGE on schemas directly to users. Instead, it grants via
a per-schema role. This is to work around the de-facto limits that various
database objects have on how many grants they can have.

We already do this for CONNECT on databases, and will follow up with doing this
for SELECT on tables in the catalogue.

Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Nick Ross <nick.ir.ross@gmail.com>