feat: find direct scheme privileges when determining available perms
Replace usage of has_schema_privilege to use aclexplode() to determine a users
direct schema permissions.

This is done primarily as a stepping stone to permissions being granted via
roles rather than directly on objects. When we move to that, we will need to
always check what permissions any role has been granted directly, rather than
walking the tree of roles, which is what has_schema_privilege does.

There should be a small performance benefit as well because roles don't need to
be checked, and PostgreSQL I think would be able to optimize a query that
doesn't use has_schema_privilege better. This is lightly verified for a single
user - the original query takes ~2.4 seconds in prod, and the new one ~1.2
seconds.

Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Nick Ross <nick.ir.ross@gmail.com>