perf: faster check on finding all existing tables when creating credentials
This change uses pg_class directly rather than the information_schema.tables view.

This reduces wall clock time from >3 seconds to < 1.5 seconds for the query.
Looking at the output of EXPLAIN ANALYZE, it also reduces the number of workers
used from 2 to 1, so in terms of total CPU time on the database, it would
reduce it from >6 seconds to 1.5 seconds.