fix: remove double quotes when fetching schema names
The ::regnamespace::text pattern wraps schemas that have dots (or other special
characters) in double quotes. Which then means they're wrapped in _another_ set
of double quotes in later queries and the first set is assumed to be part of
the name, which they're not and causes errors. We don't want the double quotes
so we can properly wrap then in a single set of double quotes later.

Potentially this isn't robust to having quotes in schema names... but that is
so unlikely.