Changing the qs limit to 50, keeping other tools at 10
Quicksight users were getting an error where there were too many connections,
so increase the limit for them. Keeping the limit for general tool users low,
as people can accidentally make thousands of connections and use all the
connections available in the database. However, this is much harder to do in
Quicksight.