feat: increase statement timeout when creating credentials
Especially when the database is under load, querying permissions on pg_class is
taking quite a while and causing tools to fail to load. We are manually
clearing up pg_class permissions separately, which should help, but to see if
can get tools working now, the best thing that can think of is to increase the
timeout.