perf: slightly faster checking for existing tables
This should make it just a touch faster to query for existing tables when
syncing permissions.

The previous version would be slower for 2 reasons:

- Querying views that each do more than they need to for our purposes.
- And the fact 3 views were queried, when 1 query on pg_class is all that's
  needed.

This is done because according to RDS performance insights, the previous query
had an every latency of ~30 seconds, which is far far too much for something
that should be really cheap.