perf: make sure to only grant roles to the master user when needed
This avoid giving the master user more roles than necessary, and only when
necessary, rather than it accumulating roles forever.

This is to try to address performance issues with the database, that from what
we can tell, are not associated with any query, but associated with starting
connections.

- In some cases just connecting to the database as the master user could take
  10 minutes.

- From the OS process list in RDS (having enabled enhanced monitoring) there
  was "start" processes from the master user with ~6% for several minutes and
  several hundres megabytes of memory. 6% is on the current setup 1 whole CPU.
  Similar start processes for non master users took much more trivial levels.

- Manually deleting the roles from the master user has resulted a tremendous
  drop of CPU usage - this PR makes is so the roles are kept down.

I suspect that this is related to the recent move to role-based permissions for
users on tables (although at first we didn't think the timelines matched up,
suspect that we were wrong there) - on connection suspect there entire "tree"
of roles would be processed in some way. With 2k roles granted to the master
user, and all of these themselves granted a few k roles, ended up with millions
of roles in the tree.