fix: don't try to revoke an app's own role when loading
This stops REVOKE of an app's own role when loading it by not including it in
the list of any "shared" roles it may have.

(Realistically an app will not have any access granted to shared roles, unless
it was done manually somehow. This code is more for human user's roles that are
granted team membership.)