refactor: make published available to new_private_database_credentials
This makes the published status of datasets available to
new_private_database_credentials.

This is a stepping stone to grouping the open datasets into a roles to
try to address connection performance problems. To do this, we need
user_access_type which is already there, but also the published status of
datasets - because superusers are allowed to access unpublished datasets. So we
are probably going to have to have 2 roles: one for published open datasets
that everyone has access to, and one for the unpublished ones that only
superusers have access to.