refactor: make a suggested role for each table available to get_new_credentials
This makes a suggested role for each table that a user has access to available
to get_new_credentials. This means that in later commits get_new_credentials
can group tables with the same suggested role, which should then reduce the
number of roles each user has, and hopefully make connecting faster.