feat: move more code to pg-sync-roles
This moves even more code to pg-sync-roles from the
new_private_database_credentials function. Primarily this is to avoid the issue
where roles would keep on accumulating on tables that are re-created a lot
(because the table select role gets granted select to the new table on each
re-creation, but then incorrectly re-created because the previous code that
this PR replaces has the table oid in the role name, and so not detected).

However, this is also to have:

- A better tested permissions system - pg-sync-roles has many tests, and it's
  relatively easy to add more
- Something with a much more concise API - now the permissions model of the
  Data Workspace database more or less fits on a single screen.
- Something that is more easily maintained by a separate team to the one that
  maintains data-workspace-frontend
- Something that can be called from elsewhere by us - for example from
  data-flow, which manages permissions for multiple database users
- Something that can be theoretically be used by others in other use cases -
  pg-sync roles is deliberately not Data Workspace-specific, and aims to be
  quite a light abstraction layer (and so, a flexible layer) over the PostgreSQL
  permissions system
- Faster sync of database permissions - the check on existing permissions in
  pg-sync-roles is much faster than the one this PR replaces, and so does not
  need caching - it does this by using pg_shdepend to find a role's existing
  permissions.

Note that the role names in the database are slightly different - pg-sync-roles
uses `_pgsr_` as a prefix. After this has been live for a few days they should
be manually cleaned up. pg-sync-roles revokes memberships not asked for, so
it's fine permissions-wise to leave the roles hanging about for a bit.